Refactored live streaming page.
Made it so that it's a bit more digestible to read and find certain content.

The accordions as a byproduct fixed a mobile overflow-x issue where the tables blew the viewport out, now they are contained to accordion's body content with overflow-x:auto;

overall performance fixes: Lazy-loading, Webp (With fallback) for lower network requests and load times.

Added a few more sections and reworded some things, and fixed typos.