streaming guide refactor (#75)
* Dark theme code block adjustment

Added `Rouge`'s class to .dark class to overwrite code blocks for a friendlier view while using dark theme.  (Global)

I tried to blend it with the overall style of the Joystick Logo.

* Spacing adjust between bullet pointer and text.

Globally adjusted the bullet to add a small gap between text and the bullet.

* Lazy load YouTube iframe.

Added lazy loading to the YouTube video to prevent unnecessary network requests until it's needed.

* asset for bad bufferbloat example.

Added webp, and jpg folder for legacy fallback using `imgsrc`.

* Rollback save.

Instead of creating a lot of static html files, using Jekylls include's and capture to insert it into the body document so that the markdown does not get screwed up.

* Removal: rogue DIV

It does not appear to serve any function, it has no class or pseudo class applied to it.

* Revert "Dark theme code block adjustment"

This reverts commit 436099fa466ea0d5835781ea8f45629e16787821.

This was changed on a different fork and merged.

* Moved img assets & created lazy-load template.

Bifurcated images into `/ext/` folder, for easy structuring. These are used for image `srcset` loading. For browsers that support webp to reduce network traffic for them, and falls back to the originals if their browser does not support it.

Created lazy_image.html to template out lazy loading for images that can use it, supports `src`, `srcset`, `alt` (Opt), `loading` (Opt) , and `caption` (Opt) for Figure caption.

* Addition and Bug fix script.js

Addition:
Jumpable anchor and expand accordion assist (e.g. `[link](#some-accordion)`

Bugfix:
Fixed the `setColorTheme` function, logic was corrected for pulling saved value from storage and comparing if it should use it.

* Added link branding to lower sidebar

* Snakeoil, sometimes I think.

dns prefetching for external assets.

* Templates and CSS for accordions

Also fixed an anchor jump issue, where it would be out of viewport when clicking sidebar ToC or content body #JumpSection in style.scss line:272

* Refactored live streaming page.

Made it so that it's a bit more digestible to read and find certain content.

The accordions as a byproduct fixed a mobile overflow-x issue where the tables blew the viewport out, now they are contained to accordion's body content with overflow-x:auto;

overall performance fixes: Lazy-loading, Webp (With fallback) for lower network requests and load times.

Added a few more sections and reworded some things, and fixed typos.

* Updated PR requests.

- Added stream from browser sections.

- Added discord link for suggestions on cloud providers.

- Reworded `Required Software` to `Streaming Software` with anchor links to mobile and desktop in-browser streaming.

* Changed JoystickTV reference to Joystick

For URL's and in the context of URL's or strings (e.g. streaming software) I left it as `joystick.tv` however for `JoystickTV` as a service I renamed it to just `Joystick` (The entity)

* Removed comment blocks that should have not been commited